+++ src/main.cpp
@@ -49,7 +49,7 @@ String CSVFileName;
 
 bool NMEAsleep = false;
 
-String nmeaData[19];
+String nmeaData[20];
 
 TaskHandle_t webTaskHandle = NULL;
 TaskHandle_t loggingTaskHandle = NULL;
@@ -97,32 +97,6 @@ String getCSV() {
     }
     return(rdata);
 }
-
-// String JSONValues() {
-//     readings["rpm"] = String(rpm);
-//     readings["etemp"] = String(etemp, 0);
-//     readings["otemp"] = String(otemp, 0);
-//     readings["opres"] = String(opres, 0);
-//     readings["fuel_rate"] = String(fuel_rate, 1);
-//     readings["flevel"] = String(flevel, 1);
-//     readings["efficiency"] = String(lpkm, 3);
-//     readings["leg_tilt"] = String(leg_tilt);
-//     readings["speed"] = String(speed);
-//     readings["heading"] = String(heading);
-//     readings["depth"] = String(depth);
-//     readings["wtemp"] = String(wtemp);
-//     readings["battV"] = String(battV);
-//     readings["ehours"] = String(ehours);
-//     readings["gear"] = String(gear);
-//     readings["lat"] = String(lat, 6);
-//     readings["lon"] = String(lon, 6);
-//     readings["mag_var"] = String(mag_var, 2);
-//     readings["time"] = String(unixTime);
-//     readings["evcErrorMsg"] = evcErrorMsg;
-//
-//     String jsonString = JSON.stringify(readings);
-//     return jsonString;
-// }
 
 bool saveSettings() {
     bool success = false;
@@ -283,9 +257,7 @@ void loop() {}
 
 void vNmeaTask(void * pvParameters) {
     TickType_t delay = 1 / portTICK_PERIOD_MS;
-    // delay = delay * 0.1;
     for (;;) {
-        // Serial.println(uxTaskGetStackHighWaterMark(NULL));
         NMEAloop();
         vTaskDelay(delay);
     }
@@ -294,7 +266,6 @@ void vNmeaTask(void * pvParameters) {
 void vWebTask(void * pvParameters) {
     for (;;) {
         webLoop();
-        // Serial.println(uxTaskGetStackHighWaterMark(NULL));
         vTaskDelay(1000 / portTICK_PERIOD_MS);
     }
 }
@@ -352,7 +323,6 @@ void vBackgroundTasks(void * pvParameters) {
             case AUTO_RPM:
                 if (nmeaData[0].toInt() <= 0) {
                     recMode=AUTO_RPM_IDLE;
-                    // endGPXfile(GPXFileName.c_str());
                     if (loggingTaskHandle != NULL) {vTaskDelete(loggingTaskHandle);}
                 }
                 localRecInt = nmeaData[0].toInt() > 3900 ? 1 : settings.recInt;
@@ -368,7 +338,6 @@ void vBackgroundTasks(void * pvParameters) {
             case AUTO_SPD:
                 if (nmeaData[8].toDouble() <= 0) {
                     recMode=AUTO_SPD_IDLE;
-                    // endGPXfile(GPXFileName.c_str());
                     if (loggingTaskHandle != NULL) {vTaskDelete(loggingTaskHandle);}
                 }
                 localRecInt = nmeaData[8].toDouble() > 15 ? 1 : settings.recInt;
@@ -390,7 +359,7 @@ void vBackgroundTasks(void * pvParameters) {
             if (outOfIdle) {
                 int voyageNum = 0;
                 String lastCSVfileName;
-                const char* csvHeaders = "RPM,Engine Temp (K),Oil Temp (K),Oil Pressure (kpa),Fuel Rate (L/h),Fuel Level (%),Fuel Efficiency (L/km),Leg Tilt (%),Speed (m/s),Heading (*),Depth (m),Water Temp (K),Battery Voltage (V),Engine Hours (s),Gear,Latitude,Longitude,Magnetic Variation (*),Time Stamp";
+                const char* csvHeaders = "RPM,Engine Temp (K),Oil Temp (K),Oil Pressure (kpa),Fuel Rate (L/h),Fuel Level (%),Fuel Efficiency (L/km),Leg Tilt (%),Speed (m/s),Heading (*),Depth (m),Water Temp (K),Battery Voltage (V),Engine Hours (s),Gear,Latitude,Longitude,Magnetic Variation (*),Time Stamp,Error Bits";
                 do {
                     voyageNum++;
                     lastCSVfileName = "Voyage";
@@ -407,7 +376,6 @@ void vBackgroundTasks(void * pvParameters) {
             count = 0;
         }
         if (NMEAsleep) {
-            // nmeaTraxGenericMsg = "N2K Bus Standby";
             if (nmeaSleepCount >= 4) {  // 5 seconds
                 nmeaSleepCount = 0;
                 NMEAsleep = false;
@@ -416,7 +384,6 @@ void vBackgroundTasks(void * pvParameters) {
                 nmeaSleepCount++;
             }
         } else {
-            // nmeaTraxGenericMsg = "";
             nmeaSleepCount = 0;
         }